Add some logs and fix the get_allowed_*() methods
This patch is adding logs to some parts of the library to help with
troubleshooting it.

The patch also fixes a problem with the get_alllowed_*() methods from
System. When the allowed values are not found in the resource JSON we
should return all values known to Sushy instead of an empty list.

Change-Id: Ifb29c85eaef47f7e112621d1c6cad4315f4c4578