Add a script to serve the Redfish mockup files
This patch is adding a small script at tools/mockup_server.py that can
be used to create a HTTP server to serve the Redfish mockup files. This
enables people to test the library without the need of having a real
hardware.

The patch also documents the use of the script.

Change-Id: I6ab21ebba27590f9eefcb7ce79620a9bc4ecec5a