Libvirt-Mockup: Get vCPUs even if the domain is not running
Prior to this patch the libvirt-mockup was relying on the libvirt's
domain.maxVcpus() method to get the number of vCPUs registered in the
domain. The problem with that method is that it only works if the domain
is running (is it a bug ? maybe). This patch will fallback and
introspect the domain XML if the domain is not running to get the number
of vCPUs from there instead.

Change-Id: Idb7a999c99592ecea74e17b67f54b48a55534f72
Closes-Bug: #1669472