Redfish-simulator: Do not cache the connection
This patch is changing the libvirt mockup script to open a new
connection with libvirt everything it needs to talk to it. Prior to it
we were caching the connection but the downside of it was that every
time the libvirt service was restarted (and it happens a couple of times
when setting up DevStack with Ironic) the simulator stopped working.

Change-Id: I94a656a91ef6d81dd2cfc52c65180beadfdcd295
Closes-Bug: #1672666