Refining the resource refresh
Introduces a public method named ``invalidate()`` in ResourceBase
and facilitates the re-fetching of the resource attributes by
invoking ``refresh()`` method only when the resource is marked as
stale. Nested resource/s are invalidated in case of resource refresh
and are lazily reloaded only when they are accessed the next time,
provided those are already initialized in contrast to recreation as
earlier.

``force`` argument introduced in ``refresh()`` method which will
force refresh the resource and its sub-resources, if set to True.
So if the resource needs to be reloaded post its initialization,
it has to be invalidated first and then refreshed:

    resource.invalidate()
    resource.refresh()

or if the resource needs to be reloaded forcefully post its
initialization, it can be achieved by:

    resource.refresh(force=True)

this will also invariably reload (greedy-refresh) its sub-resources
as well.

Co-Authored-By: Dmitry Tantsur <divius.inside@gmail.com>
Closes-Bug: 1709039

Change-Id: I89fc69ef0569c5a56abe7631d1fa6a0e5da17b80