Add system simple storage resource support
Adds the simple storage resource of Redfish standard schema.
This new resource represents the properties of a storage
controller and its directly-attached devices. This patch
introduces the property ``max_size_bytes`` of
SimpleStorageCollection resource to expose the size of the
largest storage size available among all directly attached
devices available to the System.

Also brought in the common 'Status' (comprising of Health,
HealthRollup and State sub-fields) field and refactored the code
base to use that field across all the Redfish resources.

Story: 1668487
Task: 23041

Change-Id: I512c2507bf78f4a9cf1e2525fd685836387a7581