Add system storage resource support
Adds the system storage resource of Redfish standard schema.
This new resource represents the properties of a storage subsystem
that represents a set of storage resources consisting of storage
controllers (physical or virtual), volumes and physical disk drives.
The storage volumes are built from physical disks managed by the
controllers.

This patch introduces the properties ``max_volume_size_bytes`` and
``max_drive_size_bytes`` from the StorageCollection resource to
expose the sizes of the largest storage devices (volumes and drives)
available to the System.

Story: 1668487
Task: 12506

Change-Id: If2b5114a04e2de91d3f126cd8a9af3c9c62837df