Introduce BIOS API
Add `system` resource to have reference to `bios`.

Support patching via @Redfish.Settings settings object resource
using new SettingsField.

SettingsField is a reusable field that corresponds to
@Redfish.Settings in other resources.

This SettingsField corresponds to Settings.v1_0_0 field schema, newer
versions introduce additional fields that will be added later.
Also at least iDRAC does not support anything newer than
Settings.v1_0_0 field schema now.

In the unit tests, settings.json contains imaginary values for Message,
Severity, MessageArgs and Resolution fields as could not find existing
samples holding these fields.

@Redfish.Settings is a special field (annotation) that does not appear
as a property in any schemas of Redfish resources.
Theoretically it can be present in any resource.
Practically this field is added to resources that support future state
of the resource where immediate application of updates is not possible
and need to restart system or a service for the values to take effect.
So far it has been identified that resources that can have this field
are: BIOS, Ethernet Interface and Storage.
If @Redfish.Settings is present on the resource, then need to use URI
in its subfield SettingsObject to update resource. If it is not
present on the resource, then can update the resource directly using
the resource's URI.

Change-Id: Ib2ef3d6380975ec7b964ebb634575e2b86d51991
Task: 12508
Story: 2001791