Add Redfish Message Registry resource
This is one of the first patches to support Redfish Message Registry.
It delivers only Redfish Message Registry parsing to sushy fields.
Any other processing will be added in followup patches.

Redfish Message Registry will be used by sushy internally,
as of this writing it is not intended to be exposed to sushy users.

Change-Id: I9a1735230a8328fd8365e375889c6ab066c3df16
Story: 2001791
Task: 23062