Caching ResourceCollectionBase::get_members()
Right now each time 'get_members()' method gets called a new list
is created with fresh instantiation of element type resources.

This ought to be optimized by returning the cached result in
case of repetitive invocations and should the need be to fetch
the refreshed result it would ideally reinstantiate the member
resources of the collection. Also, in case of any member resource
element within collection is being marked as stale, this handles
properly to return the refresh'ed resource if that is accessed
through this 'get_members()' method of the referenced resource
collection instance:

    res_collection_inst.get_members()[any_index].some_attr

to access the current (or live) value of 'some_attr' w/o the need
of recreating (read re-instantiating) the constituent element
resource instance.

Change-Id: I7885143baa430d4e088f99febd6f8bc1a4f99aea