Fix wrong default JsonDataReader() argument
It expect to use new JsonDataReader() instance as default argument
in function call, but it will be only created once when the function
is defined, and the same object is used in each successive call. It
does not work as expected. So Use default argument None to signal that
Reader was provided.

Change-Id: I8b659132e4699c2f9a1b8109869c3a6c1b3ca096