Fix exposed UpdateService constants
Sushy library exposed a collection of constants to the user of the
library. These constants are frequently used in library calls to
achieve desired behaviour.

The exposed constants are strings, they are designed to be
human-friendly as opposed to tedious machine identifiers.

It seems that the way UpdateService resource is implemented turns
the general principle upside down. Meaning that the constants are
taken from JSON serialization. This patch flips these things over
retaining backward compatibility.

Change-Id: I0b46db187018350062aaa1cd815fa4e029df4d3e
Story: 2003853
Task: 26652