Low case `ParamTypes` in received `MessageRegistry`
It turns out, that some implementations encountered in the wild waters
use different cases for (at least) `ParamTypes` enumerations in
`MessageRegistry` resource. In the schema [1] these enumerations
are all in low case.

To make sushy operational, this patch just ignores the case of
these enumerations. It's still an open question as to how strict the
case requirement is with Redfish (/cc Davy Jones).

1. https://redfish.dmtf.org/schemas/v1/MessageRegistry.v1_3_0.json

Change-Id: I107b689b8f4e27dd7331473124f7108d151cc55b