Add Task Monitor support for async operations
Redfish services that support asynchronous operations implement the
Task service, Task resource, and Task Monitor. When performing long
running update operations, the service returns a status of 202
Accepted. When returning a status code of 202, the response includes
a Location header containing the URL of the Task Monitor. The response
may also include the Retry-After header to specify the amount of time
the client should wait before querying the operation status.

The Task Monitor is an opaque URL that the client can query via a GET
request. While the async operation is still in progress, the service
returns a 202 Accepted status. The body of the response may be a Task
resource or it may be empty.

Once the operation completes, a GET on the Task Monitor will return a
status that indicates the result of the operation (e.g., 200, 201,
4XX).

A new TaskMonitor class is added to represent the Task Monitor and
give the sushy user the ability to monitor the status of an async
operation.

The Connector class is updated allow an option to make HTTP requests
blocking. This change utilizes the newly added TaskMonitor.

Change-Id: Ice6700806710f37b590f6b879ec38656326b39c6
Story: 2003514
Task: 30719