Handle incomplete messages in MessageRegistry
Some messges in the MessageRegistry are not having fields like
'Description' and 'Severity'. Most of the time the string in
'Message' and 'Description' fields are same. This fix proposes
to make 'Description' as not-required field and setting the
'Severity' field to default value of 'CRITICAL' if its missing
in the messages of MessageRegistry.

Example of a message in a MessageRegistry
"UpdateFailed": {
  "Message": "The update failed with a component specific error (%1).",
  "NumberOfArgs": 1,
  "ParamTypes": [
    "string"
  ],
  "Resolution": "Retry the update after remedying the component error."
}

Story: 2007216
Task: 38467
Change-Id: I3e9e2ee0867f59684c24e60772a5318801891521
(cherry picked from commit 9e12fbf109eff84ff31735a690dd010ecc4df24e)