Automatically discover available system/manager/chassis
The ``get_system``, ``get_manager`` and ``get_chassis`` methods
have been modified not to require the ``identity`` parameter
referring to a particular resource instance. If ``identity`` is
omited, sushy will default to the only available resource for as
long as it's single and therefore deterministic.

The intent is to simplify user API by not requiring the consumer
to discover available resources prior to requesting one.

Change-Id: Ifbf8b05154b619d4831c26870ee97c3dce555fd2
Story: 2007258
Task: 38707