Make MessageRegistryFile.Registry attribute non-required
According to Redfish schema, MessageRegistryFile.Registry
attribute is required and should specify some prefix for
the accompanying messages.

It turned out, that certain vendor does not supply this
attribute in Redfish response message. That makes sushy
exploding on attempt to fetch any resource from BMC.

This patch makes the above mentioned attribute defaulted
to some nonsense what will make sushy logging the problem,
but not exploding.

It is not entirely clear if the client could make sense
of the messages without this "Registry" prefix though.

Story: 2006641
Task: 36864
Change-Id: Ie8f867c7d824db156af8b874acae45d6496014a7