Adds an option for setting the http boot uri
In order to utilize Http Boot from URL, we need to be
able to set the the parameter to the remote BMC.

This enables a greater level of security as the
URL and ultimately the instruction comes from an
authenticated source, i.e. the BMC connection, and
not DHCP which is the fallback if a target is set
to UefiHttp.

We now also unset the the http boot uri if it is set
and we're not setting an explicit URI to use. This
keeps the state of the data clean in order to prevent
any unexpected behavior.

Change-Id: I73ae614ffe735bd055090ded32926a5416a12fc0