Fix OEM extension loading for different servers
Remove caching and re-using loaded extension objects
as it loaded the same instance for different servers
making it impossible to use it for more than 1
server.

The caching of extension managers is still in place,
consumer of loaded extension object can manage and
reuse the instance as needed.

Change-Id: I7834b49455258f4c3e5690708b23c67d6a361158
Story: 2007669
Task: 39767