Include extended information in debugging output
In order to help developers and operators kind of gain a
better understanding of what might be happening, we should
log the extended information.

OEM libraries extending upon sushy should be returning this
information in the errors surfaced, as applicable, however
that is not always possible so debug does seem a moderately
reasonable middle ground.

Change-Id: I3e28a3c11160fbe6fa6fe21c0f087f33d832839c