Remove auth token header completely when error occurs
While investigating a report indicating session incompatability with
Cisco BMCs, I noticed we didn't completely invalidate the session
token, and only set it to None for the request to obtain a new
token. This may not actually fix anything, but an authentication
error should completely invalidate any cached authentication data
force us through re-authentication.

Additionally, we have a bit of a lack of debug and error logging
around the session retry logic in general. Without additional
logging, we have no way to really figure out exactly what
is occuring with the BMC when we attempt to reauthenticate
the session.

Change-Id: I9eb238d1bbaf522d03ee3fce12fc0ab80c1b69b4