Make message parsing more resilient
Firmware from some vendors only includes the MessageKey in the
MessageID. In this case, fall back to finding the MessageKey first in
the Messages MessageRegistryFile and then in the BaseMessages
MessageRegistryFile.

If the message can't be found, then set the parsed message to 'unknown'
instead of throwing an exception.

If there are not enough arguments to populate the message, then
populate the missing arguments with 'unknown'.

Change-Id: I045f82167a415c92d6b67532222aef89ab1d68ef