Make Actions field in Volume resource optional
In the Volume resource implementation, the Actions field was
incorrectly specified as required. It is normal for Redfish services
to not have this field in the Volume resource. This would cause an
error when reading the Volume resource on any such Redfish service.

Updates the Actions field in the Volume resource to be optional.

Change-Id: I1399d2996d8651f6166a4909f6bdf43994ebee23
Story: 2003514
Task: 40783