Add OperationApplyTime support to Volume methods
The supported_values property in the OperationApplyTimeSupportField
class returns an unmapped list of OperationApplyTime values supported
by the service. The convention in Sushy for properties that return a
list of schema-defined enumeration values is to return a
MappedListField. This type of field performs mapping between Redfish
schema-defined enumeration values and constants exposed by the Sushy
package.

This update adds the mapped_supported_values property to return a
MappedListField of OperationApplyTime values supported by the service.
The supported_values property is deprecated.

The create_volume method in the VolumeCollection class and the
delete_volume and initialize_volume methods in the Volume class are
updated to take optional apply_time and timeout keyword parameters.
This allows the caller of those volume methods to specify a preferred
OperationApplyTime annotation and a maximum timeout for synchronous
operations. If the specified apply_time is 'Immediate', the operation
is called with blocking enabled. Otherwise blocking is disabled. For
asynchronous operations, those three methods will now return a
TaskMonitor instance that the caller can use to monitor the state of
the task.

Change-Id: I8ba2b9ff1e80fa0c2edc3a11ab80844d732e4394
Story: 2003514
Task: 41087