Fix TaskMonitor constructor calls in volume.py
In the volume.py module, the first parameter passed to the TaskMonitor
constructor was incorrect. The parameter passed was the resource object
(self), but it should have been the connector object (self._conn). This
affected the create_volume() and delete_volume() methods.

The constructor calls were updated to provide the correct connector
parameter. And the unit tests were updated to test that the TaskMonitor
objects are successfully created.

Change-Id: I10e564185b4cd9faa24739766536d39646f8a7c1
Story: 2003514
Task: 41720