Ensure Content-Type header is set when required
This used to be set to application/json when there was request data,
but this was removed in change
I6bd9a0719acfb839fcf137c58bcf03254b1af5ad without any particular
explanation.

There have been reports of some redfish implementations returning
error responses (400 or 415) when this header is not set, specifically
HPE Gen9. It is good practice to always set Content-Type on requests
with data anyway. This change adds that header back.

Change-Id: I13107bbea4d422fdfe620ade735a2e150cb51bf2