Migrate common constants to enums
The new enums are generated by a new script that reads the DMTF
definitions and produced Python code in stdout. Missing items
and docstrings are added because of that.

The change migrates everything from sushy.resources.constants except for
ProtocolType, which has to be untangled from other protocol-related
enums.

This change also misses a release note. The intention is to create one
final release note before the next release.

Change-Id: I863fad54a9703c88aa92993ad0d48d18c50e492a