Raise an AccessError with SessionService init
SessionService init previously had a check where it looked
for an AccessError exception, but actually did nothing besides
log the incident as a result.

Post-intiailization of a client, this can occur as attempts to
re-authenticate *also* force new sessions service data to be
collected and depending on session service behavior, the path
of the session session service may not be able to be retrieved,
as an access error would be raised as the refresh would include
expired session information.

Previously, there was no way to know this was happening short of an
error such as AttributError from surfacing, as this case could result
in the code short circuiting, or just ultimately the service would
log an error an the universe would attempt to move on, without any
understanding that an AccessError is occuring.

Now the AccessError is raised to the caller.

Story: 2009719
Task: 44106
Change-Id: If06fc6a52ce92753b50280a859d799f70e5799dc
(cherry picked from commit a2096cd302b343c999c6a73a0a162ac0bd8c4a23)
(cherry picked from commit d6afaf53cce774b8117110d005bf43e6456b9d8b)