Revert "Raise an AccessError with SessionService init"
This reverts commit 363d300b266d80ef55afe7ced55b87a2f41b2705.

Reason for revert: This creates a regression in fallback to guessing a path for creating session initialization, and the underlying failure was not actually causing an AttributeError to be raised but was still silently being consumed. It seems the right fix is a bit more involved and for the sake of cleanliness, reverting this patch.

Change-Id: I90def391c203fead73db6d8a4ec8261b8157c354