Revert "Raise an AccessError with SessionService init"
This reverts commit 0466e9cf5ef93088cebea518caaf2dea2e7841b9.

Reason for revert: This creates a regression in fallback to guessing a path for creating session initialization, and the underlying failure was not actually causing an AttributeError to be raised but was still silently being consumed. It seems the right fix is a bit more involved and for the sake of cleanliness, reverting this patch.

Change-Id: Icba4be519c02061e817ac7f23db1feb839d15a71