Revert "Raise an AccessError with SessionService init"
This reverts commit f435597d1fbd7d51615a4dab79c7c5d3f24cf951.

Reason for revert: This creates a regression in fallback to guessing a path for creating session initialization, and the underlying failure was not actually causing an AttributeError to be raised but was still silently being consumed. It seems the right fix is a bit more involved and for the sake of cleanliness, reverting this patch.

Change-Id: Ia34ea91ef5542b3f8c8d3cf5b201bb6e5c1430b7