Revert "Raise an AccessError with SessionService init"
This reverts commit b7a71f16ebe512d27cd00ce6f1dac90e5c4ec024.

Reason for revert: This creates a regression in fallback to guessing a path for creating session initialization, and the underlying failure was not actually causing an AttributeError to be raised but was still silently being consumed. It seems the right fix is a bit more involved and for the sake of cleanliness, reverting this patch.

Change-Id: I6b12ef536d8d26971fed790c7eee74b20f8a0b2b