Revert "Raise an AccessError with SessionService init"
This reverts commit d6afaf53cce774b8117110d005bf43e6456b9d8b.

Reason for revert: This creates a regression in fallback to guessing a path for creating session initialization, and the underlying failure was not actually causing an AttributeError to be raised but was still silently being consumed. It seems the right fix is a bit more involved and for the sake of cleanliness, reverting this patch.

Change-Id: Ib0005879c69b1cea4435ec04e0d1208e4a33f1fe