Revert "Raise an AccessError with SessionService init"
This reverts commit a2096cd302b343c999c6a73a0a162ac0bd8c4a23.

Reason for revert: This creates a regression in fallback to guessing a path for creating session initialization, and the underlying failure was not actually causing an AttributeError to be raised but was still silently being consumed. It seems the right fix is a bit more involved and for the sake of cleanliness, reverting this patch.

Change-Id: I04321b004fae5a6ae45996b83b817c2257e24f7b