Handle AccessError with Basic Auth instead of "reauth"
An operator reported issues when changing the password on a physical
machine and in Ironic, because connections are cached. However when
they supplied logs, they were actually not raising an error back to
Ironic that signaled an authentication failure, which would have
triggered the creation of a new session.

This is now handled appropriately.

Story: 2010162
Task: 45823

Change-Id: Ic52c4353cb1e1674511417223de10fa625b99af1