Do not send compression as acceptable encoding
When a BMC can compress a response, the internal server... or any
intermediate proxy may change a "Strong" ETag to a "Weak" ETag, as
the strength of the ETag represents the byte-by-byte consistency of the
response body.

Compressed responses may differ slightly, especially if different encoding
is used or negotiated. This is problematic when a resource expects a
follow-up request or action to contain a "Strong" ETag as it is the client
certifying it understood the prior precise state.

We no longer request opportunistic response encoding, and expressly request
responses are not modified or compressed by the intermediate server, which
should have a minimal impact to use and provide a higher level of BMC
interaction consistency.

Change-Id: Idb2d6bee96bd23cfc4fb8f014661231766998781
(cherry picked from commit 51dacf272f4d0f3e346c69e94786da2b7ee7c935)
(cherry picked from commit 021e0175e2ffc7c31e7052fb3bc39c9e28b83783)