Retry on ilo state error
This not great patch allows us to automatically retry when we
encounter a message from the BMC indicating that the requested
operation was invalid.

We've seen similar issues on other BMCs in the past, however
in this case the BMC uses a HTTP 400 error as opposed to a 500
error. It is not perfect, but easily accounted for with
differing approaches to the client being aware or not of the
underlying state.

Change-Id: Ib78198a60a8924de934bda0c9a0b9298541496cf
Closes-Bug: #2016307