drop usage of pkg_resource for newer python compat
pkg_resources comes from setuptools and not the python stdlib. Starting
with python 3.12 and newer, setuptools won't automatically be installed
into virtualenvs so the lack of a direct dependency will result in an
error. Further more setuptools has deprecated pkg_resources and intends
to remove it so adding a dependency would just be a band aid. The python
stdlib has added a replacement for how sushy uses pkg_resources starting
with python 3.9 and newer. Since sushy supports python 3.8 and newer we
must support using the backport package.

Depends-on: https://review.opendev.org/c/openstack/requirements/+/925306
Change-Id: I83719e368f8f7d39963c29c0013f512d1883a8dc