EthernetInterface without an Id
While the Id field is required on all Redfish object schemas, the field
can be omitted on some hardware. We do not need this field in any of the
code today so make it optional. This is no different than the Name field
being optional which is similarly required per the spec.

Closes-Bug: #2055542
Change-Id: Id78f57280261ef19dbd03046affef58546583aff
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>