enable ruff with flake8-bandit rules
Enable flake8-bandit checks via the ruff checker. The flake8-bandit
rules are prefixed with S instead of B like the bandit rules are but the
numberical values are different. flake8-bandit provides the default
config without needing it to be generated like bandit has in the ironic
repo. The only exceptions to the default config in ironic is the
skipping of 3 rules which is instead skipped just for tests instead of
globally. The choice of ruff is intentional as other OpenStack projects
are trending that way and more checks can be used in the future.

Change-Id: Iec4b628a5d5425dd18b5a4fd76f61c504e3d53b0
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>