explicit oem extn names and fix parallel tests
Rather than building up the full name and then trying to parse the
string later on, just pass the two parts to the function that needs it.
The tests are also running in parallel and the mocks and patching of the
global object falls down when there are any oem extensions installed
past the ones that match up with what the tests mock out. Currently
sushy only has one for system under contoso, which is what the first
mock uses so it just works. But adding even another causes
undeterministic failure in the tests. This should patch things and keep
them in scope to the test.

Change-Id: I8f58769b39315caf6e7939665efc9e3cc6faa0f0
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>