Handle missing storage controller mode attribute
On some Dell hardware, a storage controller may appear as a PCIe
Extender 1. These types of controllers are used internally for
connectivity to NVMe devices or other storage controllers via
the backplane.

When using such hardware, these storage controllers may not have
a controller mode attribute, which currently results in an
AttributeError.

This change adds a check for the attribute and returns None if
it is not found.

Change-Id: I25295fee0118852c4c36cec07217d720ff556587
Signed-off-by: Milan Fencik <milan.fencik@rackspace.co.uk>